+++ spacepackets/cfdp/tlv.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import dataclasses
 import struct
+from pathlib import Path
 from abc import ABC, abstractmethod
 from typing import Tuple, Optional, Type, List, Any, cast
 import enum
@@ -12,6 +13,7 @@ from spacepackets.cfdp.defs import (
     DeliveryCode,
     FileStatus,
     TransactionId,
+    TransmissionMode,
 )
 from spacepackets.exceptions import BytesTooShortError
 from spacepackets.util import UnsignedByteField
@@ -185,8 +187,8 @@ class CfdpTlv(AbstractTlvBase):
         :param value:
         :raise ValueError: Length invalid or value length not equal to specified length
         """
-        self.length = len(value)
-        if self.length > pow(2, 8) - 1:
+        self.value_len = len(value)
+        if self.value_len > pow(2, 8) - 1:
             raise ValueError("Length larger than allowed 255 bytes")
         self._tlv_type = tlv_type
         self._value = value
@@ -206,7 +208,7 @@ class CfdpTlv(AbstractTlvBase):
     def pack(self) -> bytearray:
         tlv_data = bytearray()
         tlv_data.append(self.tlv_type)
-        tlv_data.append(self.length)
+        tlv_data.append(self.value_len)
         tlv_data.extend(self._value)
         return tlv_data
 
@@ -571,7 +573,7 @@ class FileStoreRequestTlv(FileStoreRequestBase, AbstractTlvBase):
 
     def pack(self) -> bytearray:
         self.generate_tlv()
-        return self.tlv.pack()  # type: ignore
+        return self.tlv.pack()
 
     @property
     def packet_len(self):
@@ -580,7 +582,7 @@ class FileStoreRequestTlv(FileStoreRequestBase, AbstractTlvBase):
     @property
     def value(self) -> bytes:
         self.generate_tlv()
-        return self.tlv.value  # type: ignore
+        return self.tlv.value
 
     @property
     def tlv_type(self) -> TlvType:
@@ -739,11 +741,23 @@ ORIGINATING_TRANSACTION_ID_MSG_TYPE_ID = 0x0A
 
 class DirectoryOperationMessageType(enum.IntEnum):
     LISTING_REQUEST = 0x10
-    LISTING_RESPONSE = 0x10
+    LISTING_RESPONSE = 0x11
+    CUSTOM_LISTING_PARAMETERS = 0x15
+    """Custom message type not specified by the standard. Used to supply parameters like the
+    recursive and the all option to the directory listing."""
 
 
 class ReservedCfdpMessage(AbstractTlvBase):
-    """Reserved CFDP message implementation as specified in CCSDS 727.0-B-5 6.1"""
+    """Reserved CFDP message implementation as specified in CCSDS 727.0-B-5 6.1.
+
+    This class also exposes various helper types to extract the various reserved CFDP message
+    type parameters. A common way to create an instance of this class from a raw bytestream is to
+    create a :py:class:`spacepackets.cfdp.tlv.MessageToUserTlv` first after checking the TLV
+    message type, and then use the
+    :py:meth:`spacepackets.cfdp.tlv.MessageToUserTlv.is_reserved_cfdp_message`
+    and :py:meth:`spacepackets.cfdp.tlv.MessageToUserTlv.to_reserved_msg_tlv` API for the
+    conversion.
+    """
 
     def __init__(self, msg_type: int, value: bytes):
         assert msg_type < pow(2, 8) - 1
@@ -806,7 +820,7 @@ class ReservedCfdpMessage(AbstractTlvBase):
             return None
         return DirectoryOperationMessageType(self.get_reserved_cfdp_message_type())
 
-    def get_originating_transaction_id_param(
+    def get_originating_transaction_id(
         self,
     ) -> Optional[TransactionId]:
         if not self.is_originating_transaction_id():
@@ -871,6 +885,56 @@ class ReservedCfdpMessage(AbstractTlvBase):
         file_status = self.value[5] & 0b11
         return ProxyPutResponseParams(condition_code, delivery_code, file_status)
 
+    def get_proxy_closure_requested(self) -> Optional[bool]:
+        if (
+            not self.is_cfdp_proxy_operation()
+            or self.get_cfdp_proxy_message_type() != ProxyMessageType.CLOSURE_REQUEST
+        ):
+            return None
+        return self.value[5] & 0b1
+
+    def get_proxy_transmission_mode(self) -> Optional[TransmissionMode]:
+        if (
+            not self.is_cfdp_proxy_operation()
+            or self.get_cfdp_proxy_message_type() != ProxyMessageType.TRANSMISSION_MODE
+        ):
+            return None
+        return TransmissionMode(self.value[5] & 0b1)
+
+    def get_dir_listing_request_params(self) -> Optional[DirectoryParams]:
+        if (
+            not self.is_directory_operation()
+            or self.get_directory_operation_type()
+            != DirectoryOperationMessageType.LISTING_REQUEST
+        ):
+            return None
+        dir_path_lv = CfdpLv.unpack(self.value[5:])
+        dir_file_name_lv = CfdpLv.unpack(self.value[5 + dir_path_lv.packet_len :])
+        return DirectoryParams(dir_path_lv, dir_file_name_lv)
+
+    def get_dir_listing_response_params(self) -> Optional[Tuple[bool, DirectoryParams]]:
+        """
+        Returns
+        ---------
+            None if this is not a directory listing response. Otherwise, returns a tuple where
+            the first entry is a boolean denoting whether the directory listing response was
+            generated succesfully, and the second entry being the directory listing parameters.
+        """
+        if (
+            not self.is_directory_operation()
+            or self.get_directory_operation_type()
+            != DirectoryOperationMessageType.LISTING_RESPONSE
+        ):
+            return None
+        if len(self.value) < 1:
+            raise ValueError(
+                f"value with length {len(self.value)} too small for dir listing response."
+            )
+        listing_success = bool((self.value[5] >> 7) & 0b1)
+        dir_path_lv = CfdpLv.unpack(self.value[6:])
+        dir_file_name_lv = CfdpLv.unpack(self.value[6 + dir_path_lv.packet_len :])
+        return listing_success, DirectoryParams(dir_path_lv, dir_file_name_lv)
+
 
 @dataclasses.dataclass
 class ProxyPutRequestParams:
@@ -896,12 +960,31 @@ class ProxyPutResponseParams:
 
 class ProxyPutResponse(ReservedCfdpMessage):
     def __init__(self, params: ProxyPutResponseParams):
-        value = (
-            (params.condition_code << 4)
-            | (params.delivery_code << 2)
-            | params.file_status
+        super().__init__(
+            ProxyMessageType.PUT_RESPONSE,
+            bytes(
+                [
+                    (params.condition_code << 4)
+                    | (params.delivery_code << 2)
+                    | params.file_status
+                ]
+            ),
         )
-        super().__init__(ProxyMessageType.PUT_RESPONSE, value)
+
+
+class ProxyCancelRequest(ReservedCfdpMessage):
+    def __init__(self):
+        super().__init__(ProxyMessageType.PUT_CANCEL, bytes())
+
+
+class ProxyClosureRequest(ReservedCfdpMessage):
+    def __init__(self, closure_requested: bool):
+        super().__init__(ProxyMessageType.CLOSURE_REQUEST, bytes([closure_requested]))
+
+
+class ProxyTransmissionMode(ReservedCfdpMessage):
+    def __init__(self, transmission_mode: TransmissionMode):
+        super().__init__(ProxyMessageType.TRANSMISSION_MODE, bytes([transmission_mode]))
 
 
 class OriginatingTransactionId(ReservedCfdpMessage):
@@ -925,6 +1008,87 @@ class OriginatingTransactionId(ReservedCfdpMessage):
         value.extend(transaction_id.source_id.as_bytes)
         value.extend(transaction_id.seq_num.as_bytes)
         super().__init__(ORIGINATING_TRANSACTION_ID_MSG_TYPE_ID, value)
+
+
+@dataclasses.dataclass
+class DirectoryParams:
+    dir_path: CfdpLv
+    dir_file_name: CfdpLv
+
+    @classmethod
+    def from_strs(cls, dir_path: str, dir_file_name: str) -> DirectoryParams:
+        return cls(CfdpLv.from_str(dir_path), CfdpLv.from_str(dir_file_name))
+
+    @classmethod
+    def from_paths(cls, dir_path: Path, dir_file_name: Path) -> DirectoryParams:
+        return cls(CfdpLv.from_path(dir_path), CfdpLv.from_path(dir_file_name))
+
+    @property
+    def dir_path_as_str(self) -> str:
+        return self.dir_path.value.decode()
+
+    @property
+    def dir_path_as_path(self) -> Path:
+        return Path(self.dir_path_as_str)
+
+    @property
+    def dir_file_name_as_str(self) -> str:
+        return self.dir_file_name.value.decode()
+
+    @property
+    def dir_file_name_as_path(self) -> Path:
+        return Path(self.dir_file_name_as_str)
+
+
+class DirectoryListingRequest(ReservedCfdpMessage):
+    def __init__(self, params: DirectoryParams):
+        """Create a directory listing request.
+
+        Parameters
+        -----------
+
+        dir_path:
+            The path of the directory to create a listing for.
+        dir_file_name:
+            The local path at the requesting CFDP entity for the file which will contain the
+            directory listing.
+        """
+        value = params.dir_path.pack() + params.dir_file_name.pack()
+        super().__init__(DirectoryOperationMessageType.LISTING_REQUEST, value)
+
+
+class DirectoryListingResponse(ReservedCfdpMessage):
+    def __init__(self, listing_success: bool, dir_params: DirectoryParams):
+        """Create a directory listing response.
+
+        Parameters
+        -------------
+
+        listing_reponse:
+            True if the respondent CFDP user is able to proive a directory listing file.
+        dir_path:
+            The path of the directory to create a listing for.
+        dir_file_name:
+            The local path at the requesting CFDP entity for the file which will contain the
+            directory listing.
+        """
+        value = (
+            bytes([listing_success << 7])
+            + dir_params.dir_path.pack()
+            + dir_params.dir_file_name.pack()
+        )
+        super().__init__(DirectoryOperationMessageType.LISTING_RESPONSE, value)
+
+
+class DirectoryListingParameters(ReservedCfdpMessage):
+    def __init__(recursive_option: bool, all_option: bool):
+        """This is a custom reserved CFDP message to address a shortcoming of the CFDP standard
+        for directory listings.The all option could translate to something like the ``-a`` option
+        for the ``ls`` command to also display hidden files."""
+        super().__init__(
+            DirectoryOperationMessageType.CUSTOM_LISTING_PARAMETERS,
+            bytes([(recursive_option << 1) | all_option]),
+        )
 
 
 class TlvHolder: